refactor: added more type information
